requirements.txt
flask
flask-restful
flask-sqlalchemy
pymysql
flask-migrate
flask-script
flask-jwt-extended
requests
mailjet-rest
pytest
werkzeug